server/requirements.txt
Flask
PyYaml
flask-restplus
Flask-SQlAlchemy
Flask-WTF
pyodbc